requirements/detection.txt
pillow>=8.2
torch>=1.8.1
torchvision>=0.9.1
-r https://raw.githubusercontent.com/ultralytics/yolov5/master/requirements.txt